Allow to serialize/deserialize `Message`
This commit introduces `Message`, a type representing a DNS message.
One can deserialize a series of bytes into a `Message`. As well,
as deserialzing a `Message` to a series of bytes.

To support the latter, some existing types received an implenation for
`into_bytes()`.

While implementing tests to verify new code, I found a bug in
`Header::try_into()`. This bug has been fixed.