Make header API more obvious
RFC 1035 uses a lot of abbreviations to name fields and options in the
header. Initially, I used these abbreviations to name attributes in
`Header`.

However, these abbreviations are hard to understand for those less
familiar with the details of RFC 1035. Therefore, I replaced the the
abbreviations with more explicit names to make the API more obvious.