Add (de)serialize for `Name`s
Till now, this project assumed that domain names were encoded as a
series of labels. But they can also be encoded as a pointer or a series
of labels ending with a pointer.

This commit moves (de)serialization of `Name`s into it's own module and
adds support for deserializing pointers.

In order to deserialize a `Name`, I'd sometimes had to peek for the next
byte of the `Iter`. That's not possible with `Iter`, only with
`Peekable<T>`. Therefore, I'd to change the `TryFrom` implementation for
several types to `TryFrom<&mut Peekable<Iter<'_, u8>>`.